+++ client/web/src/enterprise/batches/create/CreateBatchChangePage.tsx
@@ -8,22 +8,38 @@ import { PageHeader } from '@sourcegraph/wildcard'
 import { isBatchChangesExecutionEnabled } from '../../../batches'
 import { BatchChangesIcon } from '../../../batches/icons'
 import { PageTitle } from '../../../components/PageTitle'
+import { Scalars } from '../../../graphql-operations'
 import { Settings } from '../../../schema/settings.schema'
 
-import { NewCreateBatchChangePage } from './NewCreateBatchChangePage'
+import { CreateOrEditBatchChangePage } from './CreateOrEditBatchChangePage'
 import { OldBatchChangePageContent } from './OldCreateBatchChangeContent'
 
 export interface CreateBatchChangePageProps extends SettingsCascadeProps<Settings>, ThemeProps {
+    // TODO: This can go away once we only have the new SSBC create page
     headingElement: 'h1' | 'h2'
+    /**
+     * The id for the namespace that the batch change should be created in, or that it
+     * already belongs to, if it already exists.
+     */
+    initialNamespaceID?: Scalars['ID']
 }
 
+/**
+ * CreateBatchChangePage is a wrapper around the create/edit batch change page that
+ * determines if we should display the original create page or the new SSBC page.
+ */
 export const CreateBatchChangePage: React.FunctionComponent<CreateBatchChangePageProps> = ({
     settingsCascade,
     isLightTheme,
     headingElement,
+    initialNamespaceID,
 }) =>
     isBatchChangesExecutionEnabled(settingsCascade) ? (
-        <NewCreateBatchChangePage isLightTheme={isLightTheme} settingsCascade={settingsCascade} />
+        <CreateOrEditBatchChangePage
+            isLightTheme={isLightTheme}
+            settingsCascade={settingsCascade}
+            initialNamespaceID={initialNamespaceID}
+        />
     ) : (
         <Page>
             <PageTitle title="Create batch change" />