+++ dev/sg/sg_run.go
@@ -11,6 +11,7 @@ import (
 	"github.com/peterbourgon/ff/v3/ffcli"
 
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/run"
+	"github.com/sourcegraph/sourcegraph/dev/sg/internal/stdout"
 	"github.com/sourcegraph/sourcegraph/lib/output"
 )
 
@@ -29,12 +30,12 @@ var (
 func runExec(ctx context.Context, args []string) error {
 	ok, errLine := parseConf(*configFlag, *overwriteConfigFlag)
 	if !ok {
-		out.WriteLine(errLine)
+		stdout.Out.WriteLine(errLine)
 		os.Exit(1)
 	}
 
 	if len(args) == 0 {
-		out.WriteLine(output.Linef("", output.StyleWarning, "No command specified"))
+		stdout.Out.WriteLine(output.Linef("", output.StyleWarning, "No command specified"))
 		return flag.ErrHelp
 	}
 
@@ -42,7 +43,7 @@ func runExec(ctx context.Context, args []string) error {
 	for _, arg := range args {
 		cmd, ok := globalConf.Commands[arg]
 		if !ok {
-			out.WriteLine(output.Linef("", output.StyleWarning, "ERROR: command %q not found :(", arg))
+			stdout.Out.WriteLine(output.Linef("", output.StyleWarning, "ERROR: command %q not found :(", arg))
 			return flag.ErrHelp
 		}
 		cmds = append(cmds, cmd)